+++ examples/typegoose/resolvers/types/rate-input.ts
@@ -1,0 +1,11 @@
+import { ObjectId } from "mongodb";
+import { InputType, Field, Int } from "../../../../src";
+
+@InputType()
+export class RateInput {
+  @Field()
+  recipeId: ObjectId;
+
+  @Field(type => Int)
+  value: number;
+}