+++ examples/apollo-client/src/App.tsx
@@ -1,0 +1,13 @@
+import React, { FC } from "react";
+import { ApolloProvider } from "react-apollo";
+import ApolloClient from "apollo-boost";
+
+import Counter from "./Counter/Counter";
+
+const App: FC<{ client: ApolloClient<any> }> = ({ client }) => (
+  <ApolloProvider client={client}>
+    <Counter />
+  </ApolloProvider>
+);
+
+export default App;