+++ examples/apollo-client/src/apollo/context.ts
@@ -1,0 +1,5 @@
+import { ApolloCache } from "apollo-cache";
+
+export default interface ApolloContext {
+  cache: ApolloCache<any>;
+}